requirements.txt
Flask
flask-socketio
qrcode[pil]
Pillow
python-socketio
eventlet
